+++ apps/api/src/index.ts
@@ -22,6 +22,8 @@ import { channelsRouter } from './routes/channels.js'
 import { cronRouter } from './routes/cron.js'
 import { usersRouter } from './routes/users.js'
 import { membersRouter, invitesRouter } from './routes/members.js'
+import { pluginsRouter } from './routes/plugins.js'
+
 
 import { debugRouter } from './routes/debug.js'
 import { chatRouter } from './routes/chat.js'
@@ -71,6 +73,8 @@ v1.use('/cron', cronRouter)
 v1.use('/users', usersRouter)
 v1.use('/workspaces/:id/members', membersRouter)
 v1.use('/invites', invitesRouter)
+v1.use('/plugins', pluginsRouter)
+
 v1.use('/debug', debugRouter)
 v1.use('/chat', chatRouter)